test #64
I have finally re-implemented that feature I liked so much.

When the target language is specified in the search, the 5 most frequent words in the target language are displayed. It doesn't work very well with Chinese and Japanese and in general it can still be improved a lot... But other than that it's hot stuff =]