re #51. Beginning of implementation for the sentences lists.
I have added an item called "Lists" in the menu, next to "Browse". It leads to the page that lists all the lists. From that page it is possible to create a new list.

I have added a select and a [+] button in the menu of the sentences. The <select> lists all the lists that you have created. In order to add the sentence to the list, you'll just have to select the desired list then click on [+].

It's not yet possible :
- to delete a sentence from a list
- to delete a list