re #82.
I have decided (because it's time we decide something) that the homepage will have 
- for not logged in users : what is Tatoeba, a random sentence, join Tatoeba community, number of sentences
- for logged in users : dashboard, a random sentence, latest contributions, latest comments, number of sentences

I moved the "tools" and "help" links to the bottom.
I took out the "Join Tatoeba community" (which was next to the login) to make more space. I'd rather have a big button in the corresponding module of the homepage.