re #82.
I replaced the options in the sentences menu by icons because we don't have enough space for text, and because Salem will be happier this way.

I also created a menu helper in which are defined functions that display the options in the sentences menu (because the sentence helper started to get a bit too crowded).

I took out the "Translations" <h4> thing in the sentences box, again for a matter of space. It's okay if there's only one sentence displayed, but it looks very "heavy" on the search results.