re #82.
Took out some text in the "What is Tatoeba?" and reduced margin/padding in "What can I do in Tatoeba?" to give more space for the "random sentence" (otherwise it's too low).

Rearranged Members section and user info pages.

Fixed the interface language selection (because the height was too big compared to the height of the menu bar).