Fix for bug reported here : http://forum.koohii.com/viewtopic.php?pid=86052#p86052
Quote :

Anyway, maybe you want to fix a bug with the website: if I click on "add a comment" (I am not logged in) I get the message: "You need to be logged in to add a comment. If you are not registered, you can /eng/users/register">register here."
Where "/eng/users/register">" is a link, but an invalid one.