re #113.
@milouse, the start/stop following link is working now. I fixed it because it can take you a while to figure it out and it's not worth it.

The next step would be to list the followers and followings in the user profile. At the moment they are listed in the users/show.ctp... By the way, we need to get rid of the users/show.ctp someday.

The last step would be to add another type of filter in the contribution page, which is a filter to see only the latest contributions from people we are following.