close #79.
Updated the download page and added a warning message for "normal" users to tell them that they won't find anything interesting here but that they can go check the lists section.

In the future, we can add links to download our Anki plugin, or Tatoeba Desktop, or in general, anything that is related to Tatoeba.