re #149.
Okay so this is my temporary solution until April 3rd. @Allan you can implement the tree behavior at that moment. We're not doing this for March 20th because we can't display the maintenance page yet (and I don't want to be modifying the database without displaying the maintenance page).

So what did I do. I added a parameter for the index action: index($option = null). If the parameter is null, it displays the entire wall, like it does now. If it's 'paginated' (or if it's just not null), it will display the paginated wall.

The link in the menu will redirect to the paginated version of the wall.

On the paginated version, the link to the latest messages will redirect to the message alone (because we don't know if they are on the page the user is looking at). I created a view "show_message" for that. In the future it would be more practical if this view could display the whole thread.
But in the meantime, if users want to see the message in its context, they have a link that will redirect to the non-paginated wall.

Also, the # is now redirecting tot he show_message page.

And I've created a function in the helper to display the "root" message. The messages on show_messages are displayed with it.