re #42.
Link/unlink buttons are now separated from the arrow icon. It means, if you can link or unlink, you will see two buttons. 

It's better to have a button on its own for the link/unlink feature, instead of combining it with the arrow icon. Some users might have already associated to "show", and it will confuse them if we hide a function behind it. Besides it's not a good idea to have two different functions located at the same place.