Documentation for the database.
All the scripts related to the database are now in the folder "docs/database". This folder is organized this way:

"import" => Whenever you create a new table, and this table needs to be filled with some data, you have to add in there the SQL that creates this table and fills it.

"scripts" => This folder is currently still kind of messy and not complete, but ideally there should be scripts we are likely to run regularly (for instance to clean up duplicates, to update the langStats, to update the languages fields in sentences_translations...).

"tables" => This contains the different tables of the database. They are each in a separate file. In this file, there is a description of the purpose of the table as well as a description for each field. This has to be updated every time the database changes.

"updates" => For now there is nothing, but this will be used to store scripts that need to be executed for an update. The files in this folder should be formatted like YYYYMMDD.sql, where YYYYMMDD is the date of the update. This way, all the database updates are centralized in a place where we can quickly find them, and we can keep track more easily of the changes in the database.

"database/database_20100408.sql" => It's the current structure of the database. It's just a reference.