re #211.
The sentence is not removed from the sentences_sentences_lists table, and the numberOfSentences is not updated either. Having cascading to true does not solve the issue. You have to explicitly remove the sentence from the list from the SentencesController::delete(), or from the SentenceModel::afterDelete()...

As an alternative, I have created a script that cleans up the sentences lists (deleting useless entries in sentences_sentences_lists and updating the numberOfSentences). I guess for now we can simply run this regularly, along with the duplicate deletion script.

And someday we'll set up triggers and all that fancy stuff to keep the data coherent.