re #242. Page to export into CSV.
I called it 'download' instead of export, it talks more to people (I think). There is a test() method and a test.ctp, it was just for testing how to trigger the dialog box for downloading a file. I commit this just for informative purpose, it can be deleted afterwards.

I added both the "simple" and the "anki" downloads. But for the "anki" download, I didn't give the option to choose the language of the translations yet. If we do not support Anki export for the next update, this section will either be commented out, or replaced by an explanation that it's "coming soon".