re #216.
Took out isFavorited from SentenceModel. Instead, we now check whether a sentence is favorited by the current user or not from the CurrentUserModel. 

To do so, I changed the value $useTable in FavoritedBy from 'users' into 'favorites_users'. Then created a method isSentenceFavoritedByUser(), and used this method from the CurrentUserModel.