test #236.
Using SentencesHelper::displaySentencesGroup() to display sentences in list. Now you can do everything with sentences in list.
It makes the list a too crowded in my opinion, but I will try to improve this some other time.

It would also be nice to have pagination, because displaying 500 sentences like this is not optimal...