re #348.
Simply made a page that lists sentences that don't have an owner, with the possibility to filter by language. Users can then browse these sentences and click on the "adopt" button for those they want to adopt.

Changed a little bit the text explaining what adoption is for.

Added a tip for the cases where people adopt a sentence that was already adopted shortly before.