Removed core.php and database.php config files from versioning
Instead, there is now a core.php.template and database.php.template to
copy and rename into core.php and database.php.

If there is a need to change something about the configuration, then the
corresponding template files should be modified to keep track of those
changes.

The files core.php and database.php are necessary for the app to
function and still need to be present in the config folder, but they
should never be added to version control.