Cache the search_bar element.
We only cache the 'empty' version of it, when no keywords are prefilled
and no from/to languages are selected. That is, everyone who did not search
anything so far will get a cached version. As long as a search is entered,
it stays as long as the session exists, so the performance boost isn't
that awesome.