Cache the login box element.
The form can contain prefilled values, and we don't want these
to be cached. Otherwise, everybody will get the prefilled value
someone entered at the time the cache was refreshed. Prefill
actually happens when you go on the register page and send an
invalid form. So, always keep fields empty.