#286 Removed redirection for last sentence of a certain tag. Added flash message instead.
When removing a sentence on the page tags/show_sentences_with_tag, if it
is the last sentence, we stay on the same page but display a message to
inform that the tag has been deleted.
If a user tries display a list of sentence with a tagId of tag that
doesn't exist, it will also display a message to inform that the tag
does not exist.