Fix Chosen dropdowns bug.
Chosen dropdowns are getting hidden once they get out of the
page contents’ div. This is especially annoying for dropdowns
that are on the bottom of the page, such as the ones in
/activities/translate_sentences. I’m not sure this change won’t
break other things but it sounds okay. For the record, this line
was introduced by commit b0b52ad6283f6be54786196c7ced3c4d64edf6a3.