Fix ACL bug triggered on login.
Updated User::parentNode() with code from the CakePHP 1.3 tutorial [1].

When a user logs in, we store his last_time_active by calling
User->save(). Since the User model actsAs Acl, this triggers
AclBehavior::afterSave(). But in CakePHP 1.2, it was only triggered on
user creation and not on update. Now in CakePHP 1.3, it’s triggered on
both user creation and update. This afterSave() make use of
User::parentNode() to update the Aro of the user we are working on.

The previous parentNode() code assumed that the presence of data meant
the presence of a group_id key. Which is true on user creation, but not
always true on user update. As a result, parentNode() called on the
user node was returning null instead of the group node the user
belonged to, and afterSave() was rewriting the aros table with
groupless user classes, effectively corrupting the whole ACL setup.

[1] http://book.cakephp.org/1.3/en/The-Manual/Tutorials-Examples/Simple-Acl-controlled-Application.html#acts-as-a-requester