Replace default values with jinja2-format variables
The project Tatoeba/admin contains scripts to automatically template
out cakephp configs. But these templates need to be in specific format
i.e. all the replacements must be marked by {{variable-name}} and they
will be replaced by appropriate values of variables. This commit converts
the cakephp configs in that format.