Added a script to help populating language lists.
In order to minimize the number of language names that are not translated in
the language selection dropdowns, I wrote a script to help translating
language names by using data from the CLDR project [1]. Hopefully, this will
make #327 irrelevant.

At the moment, one must download the .po from Launchpad, apply the script on
it, check if the translations are okay, and upload back the new file on
Launchpad.

[1] http://cldr.unicode.org/index/downloads