Stop ignoring diacritics, better case folding.
The previous Sphinx config was discarding all the diacritics, mapping every
character with a diacritic to its non-diacritic version. However, I think this
cause harm for searching into languages in which diacritics are as important
as if it were a different letter. For instance, in French “congrès” and
“congres” are totally different words. When one is way more frequent that the
other, sentences that include the unfrequent word gets very hard to find.

Here is a real-life example on Tatoeba: there is currently a single sentence
with the word “prisé”, and 101 sentences with the word “prise”, making that
prisé sentence very very hard to find because one has to go through all the
results to sort it out. That’s precisely because diacritics are ignored. With
this commit, looking up for “=prisé” only returns the sentence that include
that exact word. Looking up for “prisé” or “prise” (without the equal sign)
still returns sentences with close words like pris/prise/prisé…

Moreover, the case folding were broken for capital letters with diactrics and
this commit also solves that problem (#389).

I also included case folding rules for “Latin extended additional” characters
by following Unicode case character properties. I don’t know the languages
that use these characters, but again, searching with diacritics and cases
should work better for them.