Fixed bogus model relationship.
Sentences and Tags are already linked with a HABTM relation, so no
need to put it in $belongsTo. Also, CakePHP use the automatic name
'TagsSentence' for the HABTM model, which misses our 's'.

The bug appeared while loading the Sencence model *only* from the
sphinx_indexes.php shell and performing a find().