#77 Handle MySQL zeroed dates
Datetime conversion back and forth from MySQL to Python fails when datetime
is zero. MySQL allows datetime to be zero, while Python doesn't. When Python
is unable to interpret a datetime, it replaces it by None (see MySQLdb.times,
def DateTime_or_None), which stands for NULL to MySQL, whereas the column
can't be NULL, so MySQL complains.