#77 Add a user modified flag.
Change the transcription table schema.

When we update a transcription generator tool to fix things, we need to
regenerate the transcriptions. But we don’t want to mess with users by
regenerating transcriptions that has been manually modified to fix things
that may not be fixed by the generator yet. So we’ll flag manually modified
transcriptions as such, and only update the non manually modified ones.