Fix #459.
This actually fix 3 use cases:
1. Regular POST form (login, new wall post, tools…)
2. Ajax GET request (show another random sentence)
3. Ajax POST request (editing/adding sentence/translation, wall reply…)

(1) and (3) are fixed by preventing the redirection from happening in
the first place, while (2) is fixed by the null return code.

Note that somehow, CakePHP makes it like we do a PUT request when we edit
a wall post, even though a POST request was performed in the first place.