Put lang= attribute in <html>. #222
HTML5 has its own set of possible language values described in BCP 47¹.
The general rule is that you take the shortest code: 2-letters one if it
exists, 3-letters otherwise. I checked the possible values returned by
shortestCode() against the reference list², and the only values it may
return that are not in that list are: 'cycl', 'toki' and 'none'. It’s
still correct to pass unknown language attributes according to the HTML5
spec³, they are just treated as individual unknown languages by browsers.
I compiled the list using the code table from Sil⁴.

¹ http://www.ietf.org/rfc/bcp/bcp47.txt
² https://www.iana.org/assignments/language-subtag-registry
³ http://dev.w3.org/html5/spec-preview/global-attributes.html#the-lang-and-xml:lang-attributes
⁴ http://www-01.sil.org/iso639-3/download.asp
  http://www-01.sil.org/iso639-3/iso-639-3_Code_Tables_20150112.zip