Add language attrs in language dropdown elements.
Refs #222.
Since the iso codes of the interface language don’t match the ones of the
sentences, we’re not giving right codes to getLanguageDirection(). However,
it handles it correctly since the only differences are chi/cmn, fre/fra,
gre/ell and pt_BR/por, which all are LTR, on which getLanguageDirection()
defaults.