Redirect to previous page after login, not home
Closes #229.

Use GET form params instead of POST for the redirection URL,
so that we can provide it outside of forms, for instance when
clicking on sentences buttons while being logged out.

Keep the redirection URL when clicking on the top-right login link
while Javascript is disabled.

Keep the redirection URL after failed login attempts.