Refactored icons in sentences group.
This simplifies the code to display the icons in the sentences group.

Instead of displaying the icons as background, we include the SVG
inline, with <use> tag to reference to the external file
(https://css-tricks.com/svg-use-external-source/).

This allows to use `fill: currentColor` to set the color of the icon
from the CSS, which avoids having to create several SVG files for the
same icon but different colors.