Trim furiganas.
This updates the code that reads from nihongoparserd according to
allan-simon/nihongoparserd#4.

It also removes the rather useless furigana.js.
Before: we produced kanji[reading] formatted transcriptions,
        and converted them to ruby HTML tags using javascript.
Now: we use [kanji|reading] formatted transcriptions, and convert
     them to ruby HTML on server-side, using a view helper.
We need such a syntax to allow partial furiganas on tokens that start
with a kana and include a kanji, such as ご飯. With the previous syntax,
the reading はん would inevitably span on the ご too.