Only include native speakers in user group stats.
Initially the stats were counting the number of
admins/cm/ac/contributors of ANY level for each language, but I figured
this information is not that useful. We want to focus on native
speakers.

I'd like the page to give provide information about how much the
sentences in a certain language are reliable: the more native speakers
and the more the sentences are likely to be reliable since native
speakers can check, discuss, correct the sentences.