Fix /users/show
The favorites were never retrieved due to broken relationship definition.

'Sentences' isn’t a valid model name, but CakePHP seems to generate a new
model with that name automatically, which was associated with the `sentences`
table, which kinda worked.

Use 'Favorited' as related model alias to avoid name conflict with the Favorite
mode.