Add Plural-Forms gettext header parser to CakePHP
This is a backport of https://github.com/cakephp/cakephp/pull/5348
to CakePHP 1.3.

The CakePHP implementation assumes a static well-known finite list of plural
rules (in _pluralGuess()). However, not only Arabic wasn’t initially supported
(and we fixed it the hard way in 0e20cd0), but now Russian and Belarusian rules
have been updated recently on Transifex in 3b6e63b (for good, they now follows
the Unicode reference [1]), which totally breaks the original implementation.

The algorithm first converts the expression to postfix using a shunting-yard
algorithm, and then evaluates it each time a pluralized string is translated.

[1] http://www.unicode.org/cldr/charts/latest/supplemental/language_plural_rules.html#rules