Move transcription edition test to CurrentUser
So that we will be able to use it in the view.

I considered calculating a can_edit flag in the Transcription model
and pass it to the views along with the transcriptions instead (just like
the 'readonly' flag), but it wouldn’t have worked with containables,
because including 'contain' => array('Transcription') does not properly
call find() nor afterFind(), it just joins tables when doing the SQL query.