#602 Reviewed lists displayed on sentence's page.
Only display either collaborative lists or personal lists that were
created by current user.

Also reorganized display of lists: first are displayed personal lists of
the current user (the bullet is an empty circle), then are displayed the
collaborative lists (the bullet is a filled circle).

In each group (personal / collaborative), the lists are ordered by date
to which the sentence was added to the list, from the most recent to the
least recent.