Add transcriptions for Cantonese
Add an option to prevent saving anything (previous behavior).
The reason is that the current Cantonese transcription generator
directly return exponents (such as ¹²³…) that are not easy to
input at all, which makes the whole thing unusable for both manual
edition and search. We’d rather change them to regular numbers first,
but I’ll do that later.