#221 Removed change to "close" label in "reply" button of Wall messages.
It's a bit unnecessary to change the text of the "reply" to "close". Users will likely never use the "close" link on the original message to close the form.
They can use the "cancel" button in the form to close it (instead of scrolling back all the way to the original message to click on "close").