Change delta discrimination system
JSON attributes can’t be updated using UpdateAttributes(). As a result,
we need to resort on another method to get these attributes updated.
We’ll use the regular delta index instead. But for this to work, we need
a way more fine grained control to select which sentences should be
reindexed, because one attribute change on a single sentence implies
reindexation of all the translations of translations of that sentence,
since they hold information about it in their ‘trans’ attribute. So
we’re just gonna store all the sentence ids that should go into the
delta index in the new reindex_flags table.