#687 Search re-design.
I've put the name of Tatoeba above the search input and reverted the
menu item to "Home".

The "help" and "advanced search" links were moved from the bottom to the
top of the search bar.

The "from" and "to" label are displayed again.

The checkbox alignment was reverted to middle.