#731 Removed Reggiano.
Reggiano does not have an ISO 639-3 code. It's however part of the Emilian dialect, which has a code: egl.
We will therefore add the language "Emilian" instead of "Reggiano".

Currently waiting for Guybrush to send another image that we can use for the language icon, because the one he sent was meant for Reggiano only.