#805 Removed user param from lang detection query.
This causes a problem for users who have a lot of contributions and who
add a sentence in a language that is not their main language. Their
sentence is detected as written in their main language even though it's
not.

Not sure how this change will affect users with less contributions...
This might need to be reverted and the fix should be done in Tatodetect.