Remove simplified/traditional icon
It messes out with transcription toggle buttons next to it. I’m not 100% sure
it’s a good idea to drop this, but with the introduction of transcription
toggle buttons and transcription icons, one can deduce a sentence is
written in the “other” script. Anyway, we would need to get that icon
working for other scripts than Chinese at some point, if we ever show it back.