More accurate transcription warning message
The message was a bit misleading. We actually show transcriptions that are
generated but without a warning icon.

The reason a transcription may contain errors is not that it’s been
generated. It’s because the piece of software is not. As a matter of fact,
all the transcriptions are initially generated, but we consider some
already accurate enough to lack the warning sign.