Prevent normal users from reviewing others’transcr
Only allow advanced users to edit machine-generated transcription of
other members’ sentences. This is quite a sensitive right because it
can easily be used to vandalize sentences. Furthermore, I want
to allow members to contribute transcriptions in others’ sentences,
but sentence owners are not responsible of transcriptions attached
to their sentences by other members and should not be prompted to fix
them in this case. This must be fully understood by members wanting
to contribute transcriptions to others’ sentences, so let’s rescrict
this to advanced users.