Re-include ruby contents in selection
This reverts commit c1a7973c1411207cd2102b545c7f33c93df192f7.

We now have a dedicated copy-and-paste feature for the sentence
text, so there is no need to remove furigana from the selection
buffer. To the contrary, people may actually want it. If not,
they can copy the sentence text instead.

Note: it seems that newest version of Firefox automatically
strip the furigana from the copy-and-paste buffer anyway.