Furigana autogeneration optimization
DOMDocument creates a new connection to nihongoparserd for
each transcription, which adds some overhead. We get rid of
this by reusing the same connection for all the transcriptions
generated during the script execution. This is especially
useful when regenerating transcriptions in bulk using `cake
transcriptions autogen jpn`. In that case, observed a performance
boost of about 45%. I generated transcriptions for 1000 sentences
without saving them to database, and it took 5s instead of 9s.