Refactor pagination of tags
This is the first commit of a series of refacortings aiming to
get rid of getAllNeededForSentences(). Currently, when paginating
sentences, we use pagination to only get sentence ids, and then
call we that function to get data based on these sentence ids.
However, CakePHP provides a way to do this all at once using
proper pagination parameters, which is simpler and easier to
maintain. It also reduced the number of SQL queries needed to
load /tags/show_sentences_with_tag/1 (-10 queries to load
3 sentences, each having 2-3 translations).