Highlight matches when lang is unspecified, too
BuildExcerpts() doesn’t work when searching on und_main_index
(“Any”, all the indexes). In that case, we need to call
BuildExcerpts() for each language because the matching differs
according to Sphinx morphology rules. That’s why we now rely
on the sentences language instead of the queried index to guess
the index on which we should run BuildExcerpts().