Fix on the fly gen. of transcrip. for translations
This fixes on the fly generation of transcriptions of translations
(i.e. not retrieved from the database).

Now we’re using Containable to get translations, the afterFind()
method of the Transcriptable behavior doesn’t get called for
translations. So let’s call it manually instead.

This is a CakePHP limitation. Seems to be solved in 3.0.
https://github.com/cakephp/cakephp/issues/1730