Fix possible translation of unapproved sentences
Translation of unapproved sentences is only prevented by not including
the sentences.add_translation.js file for unapproved sentences.
As a result, when a page includes both unapproved and approved sentences,
the file is included and unapproved sentences can be translated normally
by clicking on their 'add translation' button, even if the tooltip says
it’s not possible.

Closes #1013.