Fix "About" page
Ref. #269: The size of the iframe was for some reason smaller than it used to, no longer applying the defined width and height. Adding the param `data-resizable="true"` seems to fix it.

Ref. #446: The params `data-initial-language` and `data-show-subtitles-default` have been added in order to have the subtitles activated by default, and displayed in the UI language if possible.

Ref. #1265: A small paragraph was added explaining what is Tatoeba and where does the name come from.

Ref. #1180: The container div is displayed with the `md-whiteframe-1dp` class.