Handle displaying of audio which author is unknown
Handle the case where both user_id is null and author is empty. This isn’t
supposed to happen because the Audio model validation rules forbid that,
but this is how we’re going to initialize the audios table when migrating.
This way, we allow old audios to be orphan, but not the new ones.