Set licence field of audio table to a string
This is more meaningful for database exports. Licence names such WTFPL
or CC BY-NC 4.0 are good enough as identifiers. We may restrict the
possible values on the CakePHP level later. As for the field author,
we will set it blank by default for old audios while making it mandatory
for new ones.