Allow more audio metadata to be stored by user
We need to save the license, along with other future metadata like
attribution URL, country/accent etc. in the User model, not Audio,
because it won’t change for all the audios of a given user, and we
want to be able to easily change this data for all the audios of a
given user at once. We’ll save the metadata as json for non-members
contributions, in order to avoid bloating the audios table with
many fields that will be only used for a very tiny portion of all
the audios (currently, 5 out of 290328).