Rename Audios controller to Audio
alanfgh said:

For what it's worth, the English word "audio" is treated as a non-count
noun, so you don't see "audios". (I realize that this is different from
"video"/"videos".) Furthermore, in a noun phrase consisting of multiple
nouns in succession, all but the last are usually in the singular (for
example, "audio list selections").