Hide dangling audio entries
Fixes #1447.

When the deduplication script merges two sentences and both of them
have audio, it leaves an entry in the `audios` table pointing to the
deleted sentence. I could fix the deduplication script, but I think
it’s better to leave it like this and just hide the dangling entries
from the listing, because these dangling entries will come handy when
we will allow multiple audio recordings per sentence. The information
they contain will make the update easier.