Prevent using license "Public domain" for audios
Giving that our sentences are currently under CC-BY, and
that audio recordings are supposedly derived work from
the textual sentences, they cannot be licensied under a
less-permissive license like Public domain.

Note that this commit only hides the option from the
browser, so it’s still possible to set "Public domain",
but this whole code is being revamped in the
sentence-licensing branch, so I’ll implement proper
validation in that branch instead.