Fix buttons after adding sentence to an empty list
In page /sentences_lists/show/<number>, when the list is empty,
the sentence JS code wasn’t loaded, because it’s only loaded
when a sentence is initially displayed on a page. However,
that page is loading sentences using AJAX queries, so we need
to have the JS loaded along with the page.

Closes #1474.