Fix setting default license
It looks like that if we set a default value on the database level,
CakePHP populates $this->data before calling beforeValidate(). This
prevents us from knowing whether those values are set by the caller
or by CakePHP, who read the schema behind our back. However, we want
be able to set a default value for the license on the model level
(because it’s a user setting), and the only way to do this is to set
NULL as default on the DB level.