Change the values of field based_on_id
It’s annoying that in SQL, "field != -1" doesn’t match records
where field is null. We need to do "field > 0 or field is null"
just to match sentences where based_on_id is known.

So let’s use these values instead:
• based_on_id = 0: sentence is original
• based_on_id > 0: sentence is based on another sentence
• based_on_id is null: we don’t know

I initially wanted to avoid having zero AND null as valid values for
that field, because it can easily lead to type casting confusions.
But it’s not very practical in SQL, so let’s just be careful with
PHP’s automatic type casting.