Be smarted when updating the number of sentences
When search is disabled, _getNumberOfSentences() will return null.
If we have some number stored in database, it means the search used
to work at some point, so let’s keep and show that number instead
of overwriting is with null. If null is stored in database, it means
the search was disabled when that vocabulary was inserted, so we just
don’t know the number.