Optimize pagination counting by using an estimate
The counting of rows in this table is terribly slow because
the table is huge. Whether we put WHERE conditions of not,
InnoDB has to scan the whole table (currently 22.5M records)
just to perform a `SELECT COUNT(*) FROM contributions`. It
looks like there isn’t much we can do about this except
using tricks like separate counting tables. But an estimate
of the row number is enough is this case, and it greatly
improve the counting speed.

Refs #1345.