Set default license from sentence author setting
Sentences may be created as belonging to someone else than the
currently logged-in user, e.g. when importing sentences, so we
need to get the default from that user’s profile instead.

Note that I had to change the type of the column 'settings',
because for some reason, when running tests, CakePHP creates
a 'users' table with a 'settings' field of type BINARY(1),
which limits its length to one character. It’s clearly not what
the doc says [1] so there is some bug here, but I just don’t to
track it down now.

[1] https://book.cakephp.org/2.0/en/development/testing.html#creating-fixtures