Fix Sphinx search with Sentence::minimalContain()
When running a search from /sentences/search while logged out,
we call find() with 'contain' => $this->Sentence->minimalContain(),
which doesn’t include the associated Transcription model. However,
addHighlightMarkers() assumed the existence of Transcription data,
and which made it adding empty strings in the $documents array
provided to BuildExcerpts(), which, for some reason, crashes the
Sphinx daemon. So everytime a search was performed, the daemon
crashed. Note that the latest version of Sphinx (3.0.3) does not
crash in this situation.