Fix testSaveTranslation_links() test
There is now a validation check on the license field when creating
a sentence. If unspecified, this field is set by default to the
value of the 'default_license' setting of the user creating the
sentence. So if the user is unspecified (as it was the case in
testSaveTranslation_links()), the license field can’t be set, and
the sentence can’t be created due to the missing license field.