Avoid locking the sentence table
Wrapping in a transaction is not really needed as we don’t
use the old deduplication script any more, so there won’t
be duplicate creation records created while we’re analysing
the contributions table. Besides, it was blocking all the
transactions attempting to update rows of the sentences
table that the script updates.