Work around stepping on one link of unrelated pair
It can happen that we look ahead after the creation log of
a sentence A and only find a single link to A, just because
we only see one link of a pair of links of a sentence B
added as a translation of A, while the other link is
outside of the maxFindAroundRange value (case 1). This can
happen no matter how big maxFindAroundRange is. In this
case, this one link is irrelevant and the sentence should
be treated as original.

At the same time, there are sentences that truly only have
one link, and for now we want to leave them with a null
based_on_id (case 2).

The way we distinguish between case 1 and case 2 is to
check the link order. In case 2, the link order is reversed
because the link is added from another sentence.