Fix test
This test was passing, but it wasn’t doing what it was
supposed to do. The call to saveTranslation() was failing
because no user was logged in. Actually, we want to
reproduce the case of the old deduplication script that
makes a reverse-ordered link pair to a sentence with an
older ID, to reach the case of count($matches) == 1 in
findLinkedSentence().