Be more careful with half-links
The last run of the script showed that none of the
sentences mentioned in the message
"Assuming sentence {$sentenceId} is original" (15 in total)
were actually original. They were all definitely added as
translation but lacking their first link for some reason.

As we want to be sure that all the sentences detected as
original really are originals, let’s ignore this special
case for the moment.