Move Sentence::$validate['license'] to __construct
so that we can use __() in the message. Note that this
wouldn’t be necessary if we were using CakePHP’s i18n
shell to extract translatable strings, because it
automatically adds them to the POT files and runs
gettext() over them, but sadly we’re using our own
script tools/generate_pot.sh instead.