Refactor ContributionTest into two tests
For easier maintenance, keeping the old test on updating the license and adding a new test for logging initial license.

Note: had to fix the old test. The values of $newLog were overriden by those of $expectedLog and the assertion was always true even if the fields were different. Made sure the array_intersect_key() is applied on the array that contains the keys we want to compare.