Re-implement search clear and language swap
Language swap needs to be re-implemented in angular because we can't use jQuery to manipulate the md-autocomplete directive.

The angular implementation is much more complicated because sadly, for reasons I couldn't figure out, it's not possible to get md-autocomplete to work inside a directive with isolated scope. It only works with `scope: true` but this prevents me from binding data between the SearchBarController and the directive. Had to resort to broadcasting events for lack of better solution.

The search clear can still work with jQuery but since we're already moving one functionality of the search bar to angular, we might as well move everything.