Implement better event architecture with listeners
I’m trying to keep this event stuff simple to use, yet not too much
complex.

About where to attach Contribution’s handlers to Sentence events:
If we do it in the controller, it won’t work with console tasks.
If we do it in Contribution::__construct(), it won’t work until we
reach a $this->Contribution kind of code because of CakePHP’s lazy
class loading. If we do it from app/Config/bootstrap.php, we
unnecessarily load models that are potentially not used at all. If
we do it from Sentence::__construct(), it works but we increase
coupling between Sentence and Contribution by specifying which
callbacks are called. So I made this thin generic listener layer
that allows to put a "Contribution listens to Sentence" kind of
generic code in Sentence::__construct(), while keeping the callbacks
within the Contribution class. Hopefully we can use this in the
future to reduce coupling in the existing code.