Make language dropdown behave more like a <select>
Display list of options on focus. I haven't found a way to programmatically pop up the list, except by clearing the search string.

We want to avoid having a value that is not a valid language in the text input. If the user doesn't select any language, we fall back to the previously selected value.

Updated the placeholder to make it clearer that a language needs to be selected, otherwise the user may be just typing the correct language name, but unless they select it, it won't work and they may be confused.