Make Queue plugin missing error easier to catch
When the Queue plugin is missing, the line:
  CakePlugin::load('Queue');
in Config/bootstrap.php throws an exception, which starts
the CakeError controller, which runs
AppController::beforeFilter(), which tries to redirect
according to $this->params['lang']. Sadly, at this time of
the loading process, the URL has not been parsed by the
Router yet, so $this->params['lang'] is undefined even
though the URL has a lang parameter. As a result, it
redirects to /eng + <URL>, which creates a redirection loop
prepending /eng over and over.

The only difference between a regular error like
MissingControllerException and that case where
$this->params['lang'] is undefined, is that the
other params like 'controller' and 'action' are
empty. That’s why we’re checking these to prevent
the redirection loop.