Fix edge case of bold menu item styling
We’re using NULL as a special value for the keys
'controller' and 'action' in $menuElements. However, it may
happen that $this->request->params['controller'] and
$this->request->params['action'] normally contain NULL:
when the parsing of parameters failed, in cases like
described in 6360880b8d27217967fa507434419c18947692ff.
In that case, everything gets styled in bold. So let’s use
false as special value instead.