Update recipient of notifications for suspended users
Not sure why but notification emails sent to community-admins@tatoeba.org (which is redirected to tatoeba-community-admins@googlegroups.com) don't work anymore.

The last notification that worked was on August 24th. Since then, the email is still being sent, but the Google Group is not receiving it... Perhaps a measure against spam.