Fix tests using Link::flagSentencesToReindex()
I guess Model::saveAll() used to call afterSave() only once,
while it now calls it for every saved record. So we need to
update our flagging strategy.

Note that the flagging was still working, it was just
adding the same sentence to flag many times, which is
not very optimized. It still does, but a little bit less.

This new code also makes valid flagging in the case of
half-linked sentences (for whatever reason it would happen).