Fix other tests of the Link model
These tests started to break because of newly-introduced
strict type checking. (in PHP array(...) == true and
array() == null).

Note that the return value of
findDirectAndIndirectTranslationsIds() is only used in

  $this->find('all', array(
    'conditions' => array('id' => $value)
  ))

kind of code, so it doesn’t matter wether it returns null
or empty array(), the find() will always return nothing.
Yet, I think it makes more sense to return an empty array
to mean "no values".