Fix redirection to last page on the Wall
Ref. #1637

Set value for $object regardless if it's an array or not. In the case of the Wall, $object is actually null and without $object being set to the model name, the rest doesn't work.

Added a test for this, but can't get it to pass. There's some error I don't have time investigate right now.