Fix test testSentenceDeletionDeletesTranscriptions
This test wasn’t working with AutoTranscriptions.enabled
set to false. Deleting transcriptions associated to a
deleted sentence has nothing to do with transcription
autogeneration, so it shouldn’t be in the Transcriptable
behavior.

Actually, we’d better do this with events and stuff,
but that’s another story for for another time.