Use "any language" as placeholder
This allows to get rid of the hack consisting of unsetting
the value to make the drop-down to open up.

It also prevents the form from sending empty values for the
'from' and 'to' parameters, which currently results in a
search error.

The language drop-down now opens up when either:
• it gets focus and "any language" is the selected value
• something is typed in or deleted in the field
• the clear field button "✘" is clicked

Using the placeholder HTML attribute as a default value is
not what it was meant for, but it improves usability a lot.

Refs #1708.