Remove 'noComment' view var
It’s useless to calculate the comment count before calling
paginate(), because paginate() will count the number of
comments too, and the case when the user exists is by far
the most frequent. In the case the user doesn’t exists,
paginate() will try to retrieve the comments, and then
not count them because there are none.