Remove crazy $persistentModel attribute
This was added in 13bb224d for performance, probably after
reading this[1]. However, according to the linked article[2],
the correct attribute name is $persistModel.

So not only this attribute had no effect, but $persistModel
was eventually removed in CakePHP 2. It’s more than time to
get rid of this voodoo.

[1] https://www.mail-archive.com/cake-php@googlegroups.com/msg65625.html
[2] https://web.archive.org/web/20090522025350/http://www.pseudocoder.com:80/archives/2009/03/17/8-ways-to-speed-up-cakephp-apps/comment-page-1/