Rework "Translate sentences" page
Ref. #1351

Removed the "Display random sentences" block. Replaced it with the options to sort by random in the search for untranslated sentences.

Restricted the languages to profile languages. There is no reason for a user to translate from/to other languages. We still keep the `none` and `und` options for the "not translated into" dropdown.

Added a hint to guide new contributors to translate into their native language in priority.