Rework SentenceTest
Split testEditSentence_fails() into several tests. We're going with one test = one use case.

Replace App::import with App::uses for Sanitize in Sentence model and remove the loading of Santize from the test class. If it's properly loaded from the model, it doesn't need to be loaded in the test.

Use assertTrue and assertFalse instead of assertEquals. It's simpler.