Port 'Revert "Use "any language" as placeholder"'
Port commit 3c52873fea0e300322d4354f76cb92311a4556e9 in the
cakephp-3 branch, so that we get rid of ' || "und" ' in the
language dropdown view. For some reasong, having the quotes
escaped with &quot; is causing a lexer error when CakePHP
spits notices before the doctype, effectively breaking all
the javascript.