Change tests on date in WallTableTest
We can't guarantee that $now will be the same minute (nor even the same year) as the saved date.

We just check if 'date' and 'modified' are equal and not null when creating a new post.
And we check that the old 'modified' value is lower than the new one when editing.