Fix editing sentences
Using CSRF component instead of middleware because I couldn't figure out how to disable CSRF on specific actions with the middleware.

We disable CSRF for the editing of sentences because it's using a jQuery plugin that doesn't provide the option to specific headers. No time to rewrite the whole edit feature right now.

Also fixed bug that would create a license log everytime sentence is edited.