Rework random sentence selection
Ref. #732

Sphinx has started to crash on selection of random sentences in any language.

Changed the random mechanism as followed: when no language defined, we first pick a random language, then we pick random sentences within this language.

The main consequence is that languages with fewer amount of sentences will have a higher chance of being displayed. Perhaps not a bad thing.