Reimplement filter by translation language
The translations are filtered and sorted by language via a custom finder function.

It's not possible to limit the language only with the query. We have to fetch at least all the direct translations, regardless of their language, so that we can check if any of the indirect translations are in the desired language.

Currently the query also fetches all the indirect translations, but we can probably optimize that to only fetch the selected translation language.