Fix localisation
The name of the folder needs to match the 2 letter code of the language (when available). CakePHP uses Locale::parseLocale() to build the path to the translation folder.

Moved the po/mo files one folder up and removed the LC_MESSAGES folders. The po/mo files need to be directly in the language folder.

Fixed typo for Portuguese in config language list.