Fix pagination for search
Added page and limit in the sphinx pagination options so we can use the proper values when calling SetLimits of the SphinxClient.

Created a finder function for sphinx search. Since the final db query only retrieves 10 entries (assuming user's settings is set at 10 results per page), we have to simulate the pagination by overriding the count and offset.