Update database scripts
Removed the acl files in the 'import' folder because they are outdated and therefore useless.

Removed the creation of tables in the script to import groups and users. The tables are already defined in the 'tables' folder. We don't want to maintain the definition in two places.

Updated the scripts in the 'acl' folder to use `bin/cake` instead of just `cake`. This way new devs can directly run `sh docs/database/acl/{file}`.