Fix license inconsitency in Terms of Use and Downloads
Our default license should be specifically CC BY 2.0 FR.

Fixed the license name on the Terms of Use which was referring to CC BY 2.0 (without FR).

Fixed the link to the CC BY license on the Downloads page (it was pointing to the generic license and not the FR one) and specified the exact version of CC BY and CC0 that we're using.